chore: migrate token over
